backend/requirements.txt
# ═══════════════════════════════════════════
# FinVerse AI — Python Dependencies
# ═══════════════════════════════════════════

# ── Core Framework ──────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1

# ── LLM Providers ──────────────────────────
google-generativeai==0.8.4
groq==0.15.0
openai==1.59.7

# ── Web Search ─────────────────────────────
tavily-python==0.5.0
google-search-results==2.4.2

# ── RAG & Embeddings ──────────────────────
sentence-transformers==3.3.1
faiss-cpu
rank-bm25==0.2.2

# ── Document Processing ───────────────────
PyPDF2==3.0.1

# ── WebSocket & Streaming ─────────────────
websockets==14.1

# ── Utilities ─────────────────────────────
httpx==0.28.1
aiofiles==24.1.0
python-multipart==0.0.20
